requirements.txt
langchain
langchain-openai
openai
python-dotenv
icecream
streamlit